+++ src/pages/profile/seller/Seller.tsx
@@ -1,12 +1,9 @@
 import Avatar from "@davatar/react";
-import dayjs from "dayjs";
-import isBetween from "dayjs/plugin/isBetween";
 import { BigNumber } from "ethers";
 import { useMemo } from "react";
 import { useParams } from "react-router-dom";
 import styled from "styled-components";
 import { useAccount } from "wagmi";
-dayjs.extend(isBetween);
 
 import {
   getLensCoverPictureUrl,
@@ -21,20 +18,18 @@ import Typography from "../../../components/ui/Typography";
 import { UrlParameters } from "../../../lib/routing/parameters";
 import { breakpoint } from "../../../lib/styles/breakpoint";
 import { colors } from "../../../lib/styles/colors";
-import { getDateTimestamp } from "../../../lib/utils/getDateTimestamp";
 import {
   MediaSet,
   Profile,
   ProfileFieldsFragment
 } from "../../../lib/utils/hooks/lens/graphql/generated";
 import useGetLensProfiles from "../../../lib/utils/hooks/lens/profile/useGetLensProfiles";
-import useProducts from "../../../lib/utils/hooks/product/useProducts";
 import { useBreakpoints } from "../../../lib/utils/hooks/useBreakpoints";
 import { useCurrentSellers } from "../../../lib/utils/hooks/useCurrentSellers";
 import { useSellerCalculations } from "../../../lib/utils/hooks/useSellerCalculations";
+import useSellerNumbers from "../../../lib/utils/hooks/useSellerNumbers";
 import { useSellers } from "../../../lib/utils/hooks/useSellers";
 import { getLensImageUrl } from "../../../lib/utils/images";
-import { ExtendedOffer } from "../../explore/WithAllOffers";
 import NotFound from "../../not-found/NotFound";
 import backgroundFluid from "../common/background-img.png";
 import ReadMore from "../common/ReadMore";
@@ -152,21 +147,7 @@ export default function Seller() {
       enabled: !!sellerId
     }
   );
-  const {
-    sellers: sellersProducts,
-    isLoading: isLoadingProducts,
-    isError: isErrorProducts
-  } = useProducts(
-    {
-      productsFilter: {
-        sellerId
-      }
-    },
-    {
-      enableCurationList: false,
-      withNumExchanges: true
-    }
-  );
+
   const {
     data: { exchanges } = {},
     isError: isErrorSellerCalculation,
@@ -180,73 +161,12 @@ export default function Seller() {
     }
   );
 
-  const products = useMemo(() => {
-    return (sellersProducts.filter((s) => s.id === sellerId) || []).map(
-      (sellerProducts) => {
-        sellerProducts.products = sellerProducts.products?.reduce(
-          (acc, elem) => {
-            const isHasBeenBought = !!elem.exchanges?.length;
-            const nowDate = dayjs();
-
-            const isFullyVoided = elem.additional?.variants.every((variant) => {
-              return !!variant.voidedAt;
-            });
-
-            // all products that are not fully voided and still have valid variants/offers
-            const isStillHaveValid =
-              !isFullyVoided &&
-              elem.additional?.variants.some((variant) => {
-                const validFromDateParsed = dayjs(
-                  Number(variant?.validFromDate) * 1000
-                );
-                const validUntilDateParsed = dayjs(
-                  Number(variant?.validUntilDate) * 1000
-                );
-                return nowDate.isBetween(
-                  validFromDateParsed,
-                  validUntilDateParsed,
-                  "day",
-                  "[]"
-                );
-              });
-
-            // all products that have been fully voided, only and only if there exist at least 1 exchange for at least 1 of the variants/offer (whatever the status of this exchange)
-            const isFullyVoidedAndBought = isFullyVoided && isHasBeenBought;
-
-            // all products where all variants are not yet valid
-            const isAllVariantsInProductNotValidYet =
-              elem.additional?.variants.every((variant) => {
-                return dayjs(getDateTimestamp(variant?.validFromDate)).isAfter(
-                  nowDate
-                );
-              });
-
-            // all products where all variants are expired, only and only if there exist at least 1 exchange for at least 1 offer (whatever the status of this exchange)
-            const isAllVariantsInProductAreExpiredAndBought =
-              elem.additional?.variants.every((variant) => {
-                return dayjs(
-                  getDateTimestamp(variant?.validUntilDate)
-                ).isBefore(nowDate);
-              }) && isHasBeenBought;
-
-            if (
-              isStillHaveValid ||
-              isFullyVoidedAndBought ||
-              isAllVariantsInProductNotValidYet ||
-              isAllVariantsInProductAreExpiredAndBought
-            ) {
-              acc.push(elem);
-            }
-            return acc;
-          },
-          [] as ExtendedOffer[]
-        );
-        // REASSIGN OFFERS for compatibility with previous code
-        sellerProducts.offers = sellerProducts.products;
-        return sellerProducts;
-      }
-    );
-  }, [sellerId, sellersProducts]);
+  const {
+    numbers: { exchanges: numExchanges, products: numProducts },
+    products,
+    isError: isErrorProducts,
+    isLoading: isLoadingProducts
+  } = useSellerNumbers(sellerId);
 
   const isSellerExists = !!sellers?.length;
   const currentSellerAddress = sellers[0]?.operator || "";
@@ -395,7 +315,7 @@ export default function Seller() {
                     margin="0"
                     fontWeight="600"
                   >
-                    {(products?.[0]?.products || [])?.length || 0}
+                    {numProducts || 0}
                   </Typography>
                 </div>
                 <div>
@@ -413,7 +333,7 @@ export default function Seller() {
                     margin="0"
                     fontWeight="600"
                   >
-                    {sellersProducts?.[0]?.numExchanges ?? 0}
+                    {numExchanges || 0}
                   </Typography>
                 </div>
                 <div>
